+++ Craftsman/Builders/Tests/UnitTests/GetEntityListUnitTestBuilder.cs
@@ -48,7 +48,7 @@ public class GetEntityListUnitTestBuilder
         var listVar = $"{entityName.LowercaseFirstLetter()}";
         var paramDto = FileNames.GetDtoName(entityName, Dto.ReadParamaters);
         var featureClassName = FileNames.GetEntityListFeatureClassName(entityName);
-        var queryListName = FileNames.QueryListName(entityName);
+        var queryListName = FileNames.QueryListName();
         
         var entityClassPath = ClassPathHelper.EntityClassPath(srcDirectory, "", entityPlural, projectBaseName);
         var dtosClassPath = ClassPathHelper.DtoClassPath(srcDirectory, "", entityPlural, projectBaseName);